Unify FFI generation into single generate_ffi.py script
Merge perform_bindgen.py and util_dyload.py into a single script that
generates both static (ffi_static.rs) and dynamic (ffi_dynamic/) FFI
files from the header file independently. Remove unused imports from
generated dynamic loading files.
