+++ dftd3/src/ffi_dynamic/dyload_compatible.rs
@@ -6,9 +6,8 @@
 //! All functions are available at runtime.
 
 use super::*;
-use core::ffi::{c_char, c_int};
 
-pub unsafe fn dftd3_get_version() -> c_int {
+pub unsafe fn dftd3_get_version() -> ::core::ffi::c_int {
     dyload_lib().dftd3_get_version.unwrap()()
 }
 
@@ -16,11 +15,15 @@ pub unsafe fn dftd3_new_error() -> dftd3_error {
     dyload_lib().dftd3_new_error.unwrap()()
 }
 
-pub unsafe fn dftd3_check_error(arg1: dftd3_error) -> c_int {
+pub unsafe fn dftd3_check_error(arg1: dftd3_error) -> ::core::ffi::c_int {
     dyload_lib().dftd3_check_error.unwrap()(arg1)
 }
 
-pub unsafe fn dftd3_get_error(arg1: dftd3_error, arg2: *mut c_char, arg3: *const c_int) {
+pub unsafe fn dftd3_get_error(
+    arg1: dftd3_error,
+    arg2: *mut ::core::ffi::c_char,
+    arg3: *const ::core::ffi::c_int,
+) {
     dyload_lib().dftd3_get_error.unwrap()(arg1, arg2, arg3)
 }
 
@@ -30,8 +33,8 @@ pub unsafe fn dftd3_delete_error(arg1: *mut dftd3_error) {
 
 pub unsafe fn dftd3_new_structure(
     arg1: dftd3_error,
-    arg2: c_int,
-    arg3: *const c_int,
+    arg2: ::core::ffi::c_int,
+    arg3: *const ::core::ffi::c_int,
     arg4: *const f64,
     arg5: *const f64,
     arg6: *const bool,
@@ -84,7 +87,7 @@ pub unsafe fn dftd3_new_zero_damping(
 
 pub unsafe fn dftd3_load_zero_damping(
     arg1: dftd3_error,
-    arg2: *mut c_char,
+    arg2: *mut ::core::ffi::c_char,
     arg3: bool,
 ) -> dftd3_param {
     dyload_lib().dftd3_load_zero_damping.unwrap()(arg1, arg2, arg3)
@@ -104,7 +107,7 @@ pub unsafe fn dftd3_new_rational_damping(
 
 pub unsafe fn dftd3_load_rational_damping(
     arg1: dftd3_error,
-    arg2: *mut c_char,
+    arg2: *mut ::core::ffi::c_char,
     arg3: bool,
 ) -> dftd3_param {
     dyload_lib().dftd3_load_rational_damping.unwrap()(arg1, arg2, arg3)
@@ -125,7 +128,7 @@ pub unsafe fn dftd3_new_mzero_damping(
 
 pub unsafe fn dftd3_load_mzero_damping(
     arg1: dftd3_error,
-    arg2: *mut c_char,
+    arg2: *mut ::core::ffi::c_char,
     arg3: bool,
 ) -> dftd3_param {
     dyload_lib().dftd3_load_mzero_damping.unwrap()(arg1, arg2, arg3)
@@ -145,7 +148,7 @@ pub unsafe fn dftd3_new_mrational_damping(
 
 pub unsafe fn dftd3_load_mrational_damping(
     arg1: dftd3_error,
-    arg2: *mut c_char,
+    arg2: *mut ::core::ffi::c_char,
     arg3: bool,
 ) -> dftd3_param {
     dyload_lib().dftd3_load_mrational_damping.unwrap()(arg1, arg2, arg3)
@@ -168,7 +171,7 @@ pub unsafe fn dftd3_new_optimizedpower_damping(
 
 pub unsafe fn dftd3_load_optimizedpower_damping(
     arg1: dftd3_error,
-    arg2: *mut c_char,
+    arg2: *mut ::core::ffi::c_char,
     arg3: bool,
 ) -> dftd3_param {
     dyload_lib().dftd3_load_optimizedpower_damping.unwrap()(arg1, arg2, arg3)
@@ -189,7 +192,7 @@ pub unsafe fn dftd3_new_cso_damping(
 
 pub unsafe fn dftd3_load_cso_damping(
     arg1: dftd3_error,
-    arg2: *mut c_char,
+    arg2: *mut ::core::ffi::c_char,
     arg3: bool,
 ) -> dftd3_param {
     dyload_lib().dftd3_load_cso_damping.unwrap()(arg1, arg2, arg3)
@@ -202,8 +205,8 @@ pub unsafe fn dftd3_delete_param(arg1: *mut dftd3_param) {
 pub unsafe fn dftd3_load_gcp_param(
     arg1: dftd3_error,
     arg2: dftd3_structure,
-    arg3: *mut c_char,
-    arg4: *mut c_char,
+    arg3: *mut ::core::ffi::c_char,
+    arg4: *mut ::core::ffi::c_char,
 ) -> dftd3_gcp {
     dyload_lib().dftd3_load_gcp_param.unwrap()(arg1, arg2, arg3, arg4)
 }
