+++ dftd3/src/ffi_dynamic/dyload_struct.rs
@@ -7,23 +7,27 @@
 //! is not found in the loaded library.
 
 use super::*;
-use core::ffi::{c_char, c_int};
 
 pub struct DyLoadLib {
     pub __libraries: Vec<libloading::Library>,
     pub __libraries_path: Vec<String>,
     pub __error: Option<String>,
-    pub dftd3_get_version: Option<unsafe extern "C" fn() -> c_int>,
+    pub dftd3_get_version: Option<unsafe extern "C" fn() -> ::core::ffi::c_int>,
     pub dftd3_new_error: Option<unsafe extern "C" fn() -> dftd3_error>,
-    pub dftd3_check_error: Option<unsafe extern "C" fn(arg1: dftd3_error) -> c_int>,
-    pub dftd3_get_error:
-        Option<unsafe extern "C" fn(arg1: dftd3_error, arg2: *mut c_char, arg3: *const c_int)>,
+    pub dftd3_check_error: Option<unsafe extern "C" fn(arg1: dftd3_error) -> ::core::ffi::c_int>,
+    pub dftd3_get_error: Option<
+        unsafe extern "C" fn(
+            arg1: dftd3_error,
+            arg2: *mut ::core::ffi::c_char,
+            arg3: *const ::core::ffi::c_int,
+        ),
+    >,
     pub dftd3_delete_error: Option<unsafe extern "C" fn(arg1: *mut dftd3_error)>,
     pub dftd3_new_structure: Option<
         unsafe extern "C" fn(
             arg1: dftd3_error,
-            arg2: c_int,
-            arg3: *const c_int,
+            arg2: ::core::ffi::c_int,
+            arg3: *const ::core::ffi::c_int,
             arg4: *const f64,
             arg5: *const f64,
             arg6: *const bool,
@@ -56,7 +60,11 @@ pub struct DyLoadLib {
         ) -> dftd3_param,
     >,
     pub dftd3_load_zero_damping: Option<
-        unsafe extern "C" fn(arg1: dftd3_error, arg2: *mut c_char, arg3: bool) -> dftd3_param,
+        unsafe extern "C" fn(
+            arg1: dftd3_error,
+            arg2: *mut ::core::ffi::c_char,
+            arg3: bool,
+        ) -> dftd3_param,
     >,
     pub dftd3_new_rational_damping: Option<
         unsafe extern "C" fn(
@@ -70,7 +78,11 @@ pub struct DyLoadLib {
         ) -> dftd3_param,
     >,
     pub dftd3_load_rational_damping: Option<
-        unsafe extern "C" fn(arg1: dftd3_error, arg2: *mut c_char, arg3: bool) -> dftd3_param,
+        unsafe extern "C" fn(
+            arg1: dftd3_error,
+            arg2: *mut ::core::ffi::c_char,
+            arg3: bool,
+        ) -> dftd3_param,
     >,
     pub dftd3_new_mzero_damping: Option<
         unsafe extern "C" fn(
@@ -85,7 +97,11 @@ pub struct DyLoadLib {
         ) -> dftd3_param,
     >,
     pub dftd3_load_mzero_damping: Option<
-        unsafe extern "C" fn(arg1: dftd3_error, arg2: *mut c_char, arg3: bool) -> dftd3_param,
+        unsafe extern "C" fn(
+            arg1: dftd3_error,
+            arg2: *mut ::core::ffi::c_char,
+            arg3: bool,
+        ) -> dftd3_param,
     >,
     pub dftd3_new_mrational_damping: Option<
         unsafe extern "C" fn(
@@ -99,7 +115,11 @@ pub struct DyLoadLib {
         ) -> dftd3_param,
     >,
     pub dftd3_load_mrational_damping: Option<
-        unsafe extern "C" fn(arg1: dftd3_error, arg2: *mut c_char, arg3: bool) -> dftd3_param,
+        unsafe extern "C" fn(
+            arg1: dftd3_error,
+            arg2: *mut ::core::ffi::c_char,
+            arg3: bool,
+        ) -> dftd3_param,
     >,
     pub dftd3_new_optimizedpower_damping: Option<
         unsafe extern "C" fn(
@@ -114,7 +134,11 @@ pub struct DyLoadLib {
         ) -> dftd3_param,
     >,
     pub dftd3_load_optimizedpower_damping: Option<
-        unsafe extern "C" fn(arg1: dftd3_error, arg2: *mut c_char, arg3: bool) -> dftd3_param,
+        unsafe extern "C" fn(
+            arg1: dftd3_error,
+            arg2: *mut ::core::ffi::c_char,
+            arg3: bool,
+        ) -> dftd3_param,
     >,
     pub dftd3_new_cso_damping: Option<
         unsafe extern "C" fn(
@@ -129,15 +153,19 @@ pub struct DyLoadLib {
         ) -> dftd3_param,
     >,
     pub dftd3_load_cso_damping: Option<
-        unsafe extern "C" fn(arg1: dftd3_error, arg2: *mut c_char, arg3: bool) -> dftd3_param,
+        unsafe extern "C" fn(
+            arg1: dftd3_error,
+            arg2: *mut ::core::ffi::c_char,
+            arg3: bool,
+        ) -> dftd3_param,
     >,
     pub dftd3_delete_param: Option<unsafe extern "C" fn(arg1: *mut dftd3_param)>,
     pub dftd3_load_gcp_param: Option<
         unsafe extern "C" fn(
             arg1: dftd3_error,
             arg2: dftd3_structure,
-            arg3: *mut c_char,
-            arg4: *mut c_char,
+            arg3: *mut ::core::ffi::c_char,
+            arg4: *mut ::core::ffi::c_char,
         ) -> dftd3_gcp,
     >,
     pub dftd3_set_gcp_realspace_cutoff:
