+++ dftd3/src/ffi_dynamic/ffi_base.rs
@@ -4,24 +4,6 @@
 
 #![allow(non_camel_case_types)]
 
-//! FFI bindings for simple-dftd3.
-//!
-//! # API Version Features
-//!
-//! This crate provides versioned FFI bindings through cargo features:
-//!
-//! - `api-v0_2`: Base API (default)
-//! - `api-v0_3`: Extends api-v0_2
-//! - `api-v0_4`: Extends api-v0_3, adds damping parameter functions
-//! - `api-v0_5`: Extends api-v0_4, adds optimized power damping and pairwise
-//!   dispersion
-//! - `api-v1_3`: Full API, adds CSO damping and GCP functions
-//!
-//! Features are cumulative: enabling `api-v0_5` also enables all functions from
-//! earlier versions (api-v0_2, api-v0_3, api-v0_4).
-
-#![allow(non_camel_case_types)]
-
 /* automatically generated by rust-bindgen 0.72.1 */
 
 #[repr(C)]
